+++ EMaster.Domain/Interfaces/EntityFramework/IGenericRepo.cs
@@ -8,7 +8,7 @@ namespace EMaster.Domain.Interfaces.EntityFramework
         TEntity AddEntity(TEntityInput entity);
         int Update(TEntity entity);
         TEntity UpdateEntity(TEntityInput entity);
-        int Delete(long id);
+        int Delete(int id);
         TEntityOutput FirstOrDefaultAsync(Expression<Func<TEntity, bool>> predicate, bool noTracking = true, params Expression<Func<TEntity, object>>[] includes);
         List<TEntityOutput> GetAll(bool noTracking = true, params string[] includes);
         TEntity FirstOrDefault(Expression<Func<TEntity, bool>> predicate, bool noTracking = true, params Expression<Func<TEntity, object>>[] includes);
